+++ build.ps1
@@ -20,6 +20,7 @@ pyinstaller --noconfirm $mode --windowed `
     --icon assets\icon.ico `
     --add-data "assets;assets" `
     --add-data "CHANGELOG.md;." `
+    --add-data "CHANGELOG.fr.md;." `
     --collect-all matplotlib `
     main.py